Update main.py
Updates:

Added a new button labeled "Generation Settings". Currently, it's non-functional, but in the future, it will allow users to modify the settings of ChatGPT to make it more creative, scientific, or accurate.
Implemented functionality to display the information entered by the user.
Established the foundation for information editing, including buttons for selecting which items the user wishes to edit.